+++ pkg/utils/kubernetes/resources/clusterrole_helpers_test.go
@@ -23,7 +23,7 @@ func TestClusterroleHelpers(t *testing.T) {
 			controlplane: "test_3.1.2",
 			image:        "kong/kubernetes-ingress-controller:3.1.2",
 			expectedClusterRole: func() *rbacv1.ClusterRole {
-				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1("test_3.1.2")
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1_lt3_2("test_3.1.2")
 				resources.LabelObjectAsControlPlaneManaged(cr)
 				return cr
 			},
@@ -33,7 +33,7 @@ func TestClusterroleHelpers(t *testing.T) {
 			image:        "kong/kubernetes-ingress-controller:3.1",
 			devMode:      true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
-				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1("test_3.1_dev")
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2("test_3.1_dev")
 				resources.LabelObjectAsControlPlaneManaged(cr)
 				return cr
 			},
@@ -48,7 +48,7 @@ func TestClusterroleHelpers(t *testing.T) {
 			image:        "kong/kubernetes-ingress-controller:3.0.0",
 			devMode:      true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
-				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1("test_3.0_dev")
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2("test_3.0_dev")
 				resources.LabelObjectAsControlPlaneManaged(cr)
 				return cr
 			},
@@ -63,7 +63,7 @@ func TestClusterroleHelpers(t *testing.T) {
 			image:        "kong/kubernetes-ingress-controller:1.0",
 			devMode:      true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
-				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1("test_unsupported_dev")
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2("test_unsupported_dev")
 				resources.LabelObjectAsControlPlaneManaged(cr)
 				return cr
 			},
@@ -78,7 +78,17 @@ func TestClusterroleHelpers(t *testing.T) {
 			image:        "test/development:main",
 			devMode:      true,
 			expectedClusterRole: func() *rbacv1.ClusterRole {
-				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_1("test_invalid_tag_dev")
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2("test_invalid_tag_dev")
+				resources.LabelObjectAsControlPlaneManaged(cr)
+				return cr
+			},
+		},
+		{
+			controlplane: "cp-3-2-0",
+			image:        "kong/kubernetes-ingress-controller:3.2.0",
+			devMode:      false,
+			expectedClusterRole: func() *rbacv1.ClusterRole {
+				cr := clusterroles.GenerateNewClusterRoleForControlPlane_ge3_2("cp-3-2-0")
 				resources.LabelObjectAsControlPlaneManaged(cr)
 				return cr
 			},