+++ pkg/utils/kubernetes/resources/clusterroles/zz_generated_controlplane_clusterrole_ge3_2.go
@@ -15,10 +15,10 @@ import (
 // ClusterRole generator
 // -----------------------------------------------------------------------------
 
-// GenerateNewClusterRoleForControlPlane_ge3_1 is a helper to generate a ClusterRole
+// GenerateNewClusterRoleForControlPlane_ge3_2 is a helper to generate a ClusterRole
 // resource with all the permissions needed by the controlplane deployment.
-// It is used for controlplanes that match the semver constraint ">=3.1"
-func GenerateNewClusterRoleForControlPlane_ge3_1(controlplaneName string) *rbacv1.ClusterRole {
+// It is used for controlplanes that match the semver constraint ">=3.2"
+func GenerateNewClusterRoleForControlPlane_ge3_2(controlplaneName string) *rbacv1.ClusterRole {
 	return &rbacv1.ClusterRole{
 		ObjectMeta: metav1.ObjectMeta{
 			GenerateName: k8sutils.TrimGenerateName(fmt.Sprintf("%s-", controlplaneName)),
@@ -178,6 +178,28 @@ func GenerateNewClusterRoleForControlPlane_ge3_1(controlplaneName string) *rbacv
 				},
 				Resources: []string{
 					"kongconsumers/status",
+				},
+				Verbs: []string{
+					"get", "patch", "update",
+				},
+			},
+			{
+				APIGroups: []string{
+					"configuration.konghq.com",
+				},
+				Resources: []string{
+					"kongcustomentities",
+				},
+				Verbs: []string{
+					"get", "list", "watch",
+				},
+			},
+			{
+				APIGroups: []string{
+					"configuration.konghq.com",
+				},
+				Resources: []string{
+					"kongcustomentities/status",
 				},
 				Verbs: []string{
 					"get", "patch", "update",